+++ include/odp/arch/arm64-linux/odp/api/abi/cpu.h
@@ -13,16 +13,6 @@ extern "C" {
 
 #define ODP_CACHE_LINE_SIZE 64
 
-static inline void odp_cpu_pause(void)
-{
-	/* YIELD hints the CPU to switch to another thread if possible
-	 * and executes as a NOP otherwise.
-	 * ISB flushes the pipeline, then restarts. This is guaranteed to
-	 * stall the CPU a number of cycles.
-	 */
-	__asm volatile("isb" ::: "memory");
-}
-
 #ifdef __cplusplus
 }
 #endif
