+++ include/odp/arch/default-linux/odp/api/abi/cpu.h
@@ -13,10 +13,6 @@ extern "C" {
 
 #define ODP_CACHE_LINE_SIZE 64
 
-static inline void odp_cpu_pause(void)
-{
-}
-
 #ifdef __cplusplus
 }
 #endif
