+++ include/odp/arch/x86_64-linux/odp/api/abi/cpu.h
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier:     BSD-3-Clause
  */
 
-#ifndef ODP_API_ABI_CPU_ARCH_H_
-#define ODP_API_ABI_CPU_ARCH_H_
+#ifndef ODP_API_ABI_CPU_H_
+#define ODP_API_ABI_CPU_H_
 
 #ifdef __cplusplus
 extern "C" {
